Remove deoplete-omnisharp
This was almost the oldest one around

https://gitlab.com/mixedCase/deoplete-omnisharp

and apparently almost dramatic

https://github.com/Shougo/deoplete.nvim/issues/217

*and* seems it was the `CS` source when there were `OS`
sources as well.

Now `OmniSharp-vim` says it uses the stdio server instead
of the HTTP server, so maybe this is just safe to remove.

I don't have any actual Unity code on this laptop, but
looking at `spine-csharp`, I get encouraging results!

As long as the .sln file is available, that is...